fix match history page
